--- VendingException.java
@@ -1,7 +1,0 @@
-public class VendingException extends RuntimeException
-{
-   public VendingException(String reason)
-   {
-      super(reason);
-   }
-}
