+++ coin/Coin.java
