+++ coin/CoinLineItem.class
