+++ coin/CoinLineItem.java
