+++ coin/CoinSet.class
