+++ txt/Operators.txt
