+++ vendingMachine/VendingException.java
@@ -1,0 +1,7 @@
+public class VendingException extends RuntimeException
+{
+   public VendingException(String reason)
+   {
+      super(reason);
+   }
+}
