+++ vendingMachine/VendingMachineSimulation.class
