+++ vendingMachine/VendingMachineSimulation.java
